Add ORM configuration, slightly change project structure
