+++ backend/.env
@@ -1,6 +1,8 @@
 PORT=8001
 DB_HOST="localhost"
+DB_PORT=3306
 DB_USER="root"
 DB_PWD=""
 DB_NAME="db"
+DB_SYNC=true
 CORS_ORIGIN=localhost:8080
