--- backend/config/db.config.ts
@@ -1,10 +1,0 @@
-import mysql from "mysql2";
-import * as dotenv from "dotenv";
-dotenv.config();
-
-export const db = mysql.createConnection({
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PWD,
-  database: process.env.DB_NAME,
-});
