+++ backend/package.json
@@ -2,11 +2,11 @@
   "name": "watercool-backend",
   "version": "1.0.0",
   "description": "",
-  "main": "server.ts",
+  "main": "index.ts",
   "scripts": {
     "test": "npm run test",
-    "start": "ts-node server.ts",
-    "dev": "nodemon --exec \"ts-node\"  server.ts",
+    "start": "ts-node src/index.ts",
+    "dev": "nodemon --exec \\\"ts-node\\\"  src/index.ts",
     "build": "tsc && node server.js"
   },
   "repository": {
@@ -24,7 +24,9 @@
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "mysql2": "^2.2.5"
+    "mysql2": "^2.2.5",
+    "reflect-metadata": "^0.1.13",
+    "typeorm": "^0.2.36"
   },
   "devDependencies": {
     "@types/body-parser": "^1.19.1",
